Fixed issue where hh notes would not load...
When selecting a hunting horn after a weapon that has less detail rows,
the rows with notes would not load properly, depending on the amount of rows;
the notes would either not load or load partially or they would load properly.
To solve it I added a manual resize to force a repaint since other methods failed.